Normalize PostgreSQL SSL config for runtime DB access
